refactor(bienbot-ui): extract SessionSharePopover
Move the inline SessionSharePopover component (~127 lines) from
BienBotPanel.jsx into its own file for easier maintenance and testing.
